feat!: implement orderbook v2 core

+++ src/interfaces/IConditionalTokens.sol
@@ -13,18 +13,15 @@ interface IConditionalTokens is IERC1155 {
     /// @param questionId An identifier for the question to be answered by the oracle.
     /// @param outcomeSlotCount The number of outcome slots which should be used for this condition. Must not exceed 256.
     event ConditionPreparation(
-        bytes32 indexed conditionId,
-        address indexed oracle,
-        bytes32 indexed questionId,
-        uint outcomeSlotCount
+        bytes32 indexed conditionId, address indexed oracle, bytes32 indexed questionId, uint256 outcomeSlotCount
     );
 
     event ConditionResolution(
         bytes32 indexed conditionId,
         address indexed oracle,
         bytes32 indexed questionId,
-        uint outcomeSlotCount,
-        uint[] payoutNumerators
+        uint256 outcomeSlotCount,
+        uint256[] payoutNumerators
     );
 
     /// @dev Emitted when a position is successfully split.
@@ -33,8 +30,8 @@ interface IConditionalTokens is IERC1155 {
         IERC20 collateralToken,
         bytes32 indexed parentCollectionId,
         bytes32 indexed conditionId,
-        uint[] partition,
-        uint amount
+        uint256[] partition,
+        uint256 amount
     );
     /// @dev Emitted when positions are successfully merged.
     event PositionsMerge(
@@ -42,16 +39,16 @@ interface IConditionalTokens is IERC1155 {
         IERC20 collateralToken,
         bytes32 indexed parentCollectionId,
         bytes32 indexed conditionId,
-        uint[] partition,
-        uint amount
+        uint256[] partition,
+        uint256 amount
     );
     event PayoutRedemption(
         address indexed redeemer,
         IERC20 indexed collateralToken,
         bytes32 indexed parentCollectionId,
         bytes32 conditionId,
-        uint[] indexSets,
-        uint payout
+        uint256[] indexSets,
+        uint256 payout
     );
 
     /// Mapping key is an condition ID. Value represents numerators of the payout vector associated with the condition. This array is initialized with a length equal to the outcome slot count. E.g. Condition with 3 outcomes [A, B, C] and two of those correct [0.5, 0.5, 0]. In Ethereum there are no decimal values, so here, 0.5 is represented by fractions like 1/2 == 0.5. That's why we need numerator and denominator values. Payout numerators are also used as a check of initialization. If the numerators array is empty (has length zero), the condition was not created/prepared. See getOutcomeSlotCount.
@@ -64,12 +61,12 @@ interface IConditionalTokens is IERC1155 {
     /// @param oracle The account assigned to report the result for the prepared condition.
     /// @param questionId An identifier for the question to be answered by the oracle.
     /// @param outcomeSlotCount The number of outcome slots which should be used for this condition. Must not exceed 256.
-    function prepareCondition(address oracle, bytes32 questionId, uint outcomeSlotCount) external;
+    function prepareCondition(address oracle, bytes32 questionId, uint256 outcomeSlotCount) external;
 
     /// @dev Called by the oracle for reporting results of conditions. Will set the payout vector for the condition with the ID ``keccak256(abi.encodePacked(oracle, questionId, outcomeSlotCount))``, where oracle is the message sender, questionId is one of the parameters of this function, and outcomeSlotCount is the length of the payouts parameter, which contains the payoutNumerators for each outcome slot of the condition.
     /// @param questionId The question ID the oracle is answering for
     /// @param payouts The oracle's answer
-    function reportPayouts(bytes32 questionId, uint[] calldata payouts) external;
+    function reportPayouts(bytes32 questionId, uint256[] calldata payouts) external;
 
     /// @dev This function splits a position. If splitting from the collateral, this contract will attempt to transfer `amount` collateral from the message sender to itself. Otherwise, this contract will burn `amount` stake held by the message sender in the position being split worth of EIP 1155 tokens. Regardless, if successful, `amount` stake will be minted in the split target positions. If any of the transfers, mints, or burns fail, the transaction will revert. The transaction will also revert if the given partition is trivial, invalid, or refers to more slots than the condition is prepared with.
     /// @param collateralToken The address of the positions' backing collateral token.
@@ -81,8 +78,8 @@ interface IConditionalTokens is IERC1155 {
         IERC20 collateralToken,
         bytes32 parentCollectionId,
         bytes32 conditionId,
-        uint[] calldata partition,
-        uint amount
+        uint256[] calldata partition,
+        uint256 amount
     ) external;
 
     /// @dev This function merges CTF tokens into the underlying collateral.
@@ -95,8 +92,8 @@ interface IConditionalTokens is IERC1155 {
         IERC20 collateralToken,
         bytes32 parentCollectionId,
         bytes32 conditionId,
-        uint[] calldata partition,
-        uint amount
+        uint256[] calldata partition,
+        uint256 amount
     ) external;
 
     /// @dev This function redeems a CTF ERC1155 token for the underlying collateral
@@ -104,27 +101,38 @@ interface IConditionalTokens is IERC1155 {
     /// @param parentCollectionId The ID of the outcome collections common to the position
     /// @param conditionId The ID of the condition to split on.
     /// @param indexSets Index sets of the outcome collection to combine with the parent outcome collection
-    function redeemPositions(IERC20 collateralToken, bytes32 parentCollectionId, bytes32 conditionId, uint[] calldata indexSets) external;
+    function redeemPositions(
+        IERC20 collateralToken,
+        bytes32 parentCollectionId,
+        bytes32 conditionId,
+        uint256[] calldata indexSets
+    ) external;
 
     /// @dev Gets the outcome slot count of a condition.
     /// @param conditionId ID of the condition.
     /// @return Number of outcome slots associated with a condition, or zero if condition has not been prepared yet.
-    function getOutcomeSlotCount(bytes32 conditionId) external view returns (uint);
+    function getOutcomeSlotCount(bytes32 conditionId) external view returns (uint256);
 
     /// @dev Constructs a condition ID from an oracle, a question ID, and the outcome slot count for the question.
     /// @param oracle The account assigned to report the result for the prepared condition.
     /// @param questionId An identifier for the question to be answered by the oracle.
     /// @param outcomeSlotCount The number of outcome slots which should be used for this condition. Must not exceed 256.
-    function getConditionId(address oracle, bytes32 questionId, uint outcomeSlotCount) external pure returns (bytes32);
+    function getConditionId(address oracle, bytes32 questionId, uint256 outcomeSlotCount)
+        external
+        pure
+        returns (bytes32);
 
     /// @dev Constructs an outcome collection ID from a parent collection and an outcome collection.
     /// @param parentCollectionId Collection ID of the parent outcome collection, or bytes32(0) if there's no parent.
     /// @param conditionId Condition ID of the outcome collection to combine with the parent outcome collection.
     /// @param indexSet Index set of the outcome collection to combine with the parent outcome collection.
-    function getCollectionId(bytes32 parentCollectionId, bytes32 conditionId, uint indexSet) external view returns (bytes32);
+    function getCollectionId(bytes32 parentCollectionId, bytes32 conditionId, uint256 indexSet)
+        external
+        view
+        returns (bytes32);
 
     /// @dev Constructs a position ID from a collateral token and an outcome collection. These IDs are used as the ERC-1155 ID for this contract.
     /// @param collateralToken Collateral token which backs the position.
     /// @param collectionId ID of the outcome collection associated with this position.
-    function getPositionId(IERC20 collateralToken, bytes32 collectionId) external pure returns (uint);
+    function getPositionId(IERC20 collateralToken, bytes32 collectionId) external pure returns (uint256);
 }
